// web/src/app/shared/api-client/model/flashcard-content-text-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FlashcardContentDto } from './flashcard-content-dto';


export interface FlashcardContentTextDto extends FlashcardContentDto { 
    text: string;
    languageId: string;
}