// web/src/app/shared/api-client/model/flashcard-content-upload-image-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FlashcardContentDto } from './flashcard-content-dto';


export interface FlashcardContentUploadImageDto extends FlashcardContentDto { 
    imageFileName?: string;
}